Remove mssql mock, unify SQL-based query tests.
